PatternCreator function moved to helpers/PatternCreator for less clutter
